+++ python/aistore/sdk/const.py
@@ -39,13 +39,17 @@ AIS_OBJ_NAME = "ais-name"
 AIS_LOCATION = "ais-location"
 AIS_MIRROR_PATHS = "ais-mirror-paths"
 AIS_MIRROR_COPIES = "ais-mirror-copies"
+AIS_EC_GENERATION = "ais-ec-generation"
+AIS_EC_DATA = "ais-ec-data"
+AIS_EC_PARITY = "ais-ec-parity"
+AIS_EC_REPLICATED = "ais-ec-replicated"
 AIS_PRESENT = "ais-present"
-# V2-specific headers
+# Selective HEAD headers
 AIS_CHUNKS_COUNT = "ais-chunks-count"
 AIS_CHUNKS_MAX_CHUNK_SIZE = "ais-chunks-max-chunk-size"
 HEADER_LAST_MODIFIED = "Last-Modified"
 HEADER_ETAG = "ETag"
-# V2 property names
+# Selective HEAD property names
 PROPS_CHUNKED = "chunked"
 # Bucket Props Header keys
 HEADER_PREFIX = "Ais-"
@@ -96,7 +100,7 @@ QPARAM_NEW_CUSTOM = "set-new-custom"
 # multipart upload
 QPARAM_MPT_UPLOAD_ID = "uploadId"
 QPARAM_MPT_PART_NO = "partNumber"
-# Query params for V2 HEAD
+# Query params for selective object HEAD
 QPARAM_PROPS = "props"
 # etl
 QPARAM_ETL_NAME = "etl_name"