+++ python/tests/unit/sdk/obj/test_object.py
@@ -37,14 +37,12 @@ from aistore.sdk.const import (
     HEADER_OBJECT_BLOB_DOWNLOAD,
     HEADER_OBJECT_BLOB_CHUNK_SIZE,
     HEADER_OBJECT_BLOB_WORKERS,
-    AIS_BCK_NAME,
-    AIS_BCK_PROVIDER,
-    AIS_OBJ_NAME,
     AIS_LOCATION,
     AIS_MIRROR_PATHS,
     AIS_MIRROR_COPIES,
     AIS_PRESENT,
     QPARAM_LATEST,
+    QPARAM_PROPS,
     QPARAM_SYNC,
 )
 from aistore.sdk.obj.object import (
@@ -83,6 +81,7 @@ class TestObject(unittest.TestCase):
         )
         self.mock_writer = Mock()
         self.expected_params = self.bck_qparams
+        self.mock_client.request.return_value = Mock(headers=CaseInsensitiveDict())
         self.object = Object(self.mock_client, self.bucket_details, OBJ_NAME)
 
     def test_properties(self):
@@ -94,13 +93,18 @@ class TestObject(unittest.TestCase):
         self.assertIsInstance(self.object.props, ObjectProps)
 
     def test_head(self):
-        self.object.head()
+        self.assertFalse(hasattr(self.object, "head_v2"))
+        attributes = self.object.head()
 
         self.mock_client.request.assert_called_with(
             HTTP_METHOD_HEAD,
             path=REQUEST_PATH,
-            params=self.expected_params,
+            params={**self.expected_params, QPARAM_PROPS: "name,size"},
         )
+        self.assertIs(attributes, self.object.props_cached)
+        self.assertEqual(BCK_NAME, attributes.bucket_name)
+        self.assertEqual("ais", attributes.bucket_provider)
+        self.assertEqual(OBJ_NAME, attributes.name)
 
     def test_get_default_params(self):
         self.get_exec_assert()
@@ -414,14 +418,11 @@ class TestObject(unittest.TestCase):
         headers = CaseInsensitiveDict(
             {
                 "Ais-Atime": "1722021816727999173",
-                "Ais-Bucket-Name": "data-bck",
-                "Ais-Bucket-Provider": "ais",
                 "Ais-Checksum-Type": "xxhash",
                 "Ais-Checksum-Value": "ecc0a7bf787e089e",
                 "Ais-Location": "t[LSJt8081]:mp[/tmp/ais/mp1/1, [sda sdb]]",
                 "Ais-Mirror-Copies": "1",
                 "Ais-Mirror-Paths": "[/tmp/ais/mp1/1]",
-                "Ais-Name": "cifar-10-batches-py/batches.meta",
                 "Ais-Present": "true",
                 "Ais-Version": "1",
                 "Content-Length": "158",
@@ -433,13 +434,20 @@ class TestObject(unittest.TestCase):
 
         self.assertEqual(self.object.props_cached, None)
 
-        self.object.head()
-
         props: ObjectProps = self.object.props
 
-        self.assertEqual(props.bucket_name, headers[AIS_BCK_NAME])
-        self.assertEqual(props.bucket_provider, headers[AIS_BCK_PROVIDER])
-        self.assertEqual(props.name, headers[AIS_OBJ_NAME])
+        self.mock_client.request.assert_called_with(
+            HTTP_METHOD_HEAD,
+            path=REQUEST_PATH,
+            params={
+                **self.expected_params,
+                QPARAM_PROPS: "checksum,atime,version,copies,custom,location",
+            },
+        )
+
+        self.assertEqual(props.bucket_name, BCK_NAME)
+        self.assertEqual(props.bucket_provider, "ais")
+        self.assertEqual(props.name, OBJ_NAME)
         self.assertEqual(props.location, headers[AIS_LOCATION])
         self.assertEqual(
             props.mirror_paths, headers[AIS_MIRROR_PATHS].strip("[]").split(",")